# Conflicts:
#	Biovation.Brands.Suprema/Commands/Suprema_GetLogsOfDeviceInPeriod.cs
#	Biovation.Brands.Suprema/Commands/Suprema_RetrieveUserFromDevice.cs
#	Biovation.Brands.Suprema/Commands/Suprema_SyncUserOfDevice.cs
#	Biovation.Brands.Suprema/Commands/Suprema_SyncUsersOfDevice.cs
#	Biovation.Brands.Suprema/Devices/Suprema Version 1/OtherDevices.cs
#	Biovation.Brands.Suprema/Manager/TaskManager.cs
#	Biovation.Service.Api/v1/FingerTemplateService.cs

+++ Biovation.Domain/User.cs
@@ -142,7 +142,7 @@ namespace Biovation.Domain
             return Convert.ToInt32(StartDate.Ticks / 10000000 - refDate);
         }
 
-        public void SetStartDateFromTicks(int startDateTicks)
+        public void SetStartDateFromTicks(long startDateTicks)
         {
             StartDate = new DateTime(1970, 1, 1).AddTicks(startDateTicks);
         }
@@ -153,7 +153,7 @@ namespace Biovation.Domain
             return Convert.ToInt32(EndDate.Ticks / 10000000 - refDate);
         }
 
-        public void SetEndDateFromTicks(int endDateTicks)
+        public void SetEndDateFromTicks(long endDateTicks)
         {
             EndDate = new DateTime(1970, 1, 1).AddTicks(endDateTicks);
         }
